requirements.txt
fastapi
torch
torchvision
transformers
huggingface_hub
peft
Pillow
